refactor(anyharness): materialization runtime valve (grid PR 7)
Gives the materialization domain its runtime valve, mirroring the mobility
(6a) shape: MaterializationService keeps durable/pure logic (repo-root
acquisition, ledger admission), while a new MaterializationRuntime owns the
one live-touching use case (exact-ref workspace materialization's reuse
admission check, which asks live sessions and live terminals whether a
reused workspace is busy). The service and runtime share one ledger store
and one in-process lock map (constructed once in app/materialization.rs)
since repo-root and workspace operation ids live in the same namespace and
must converge/conflict against the same state regardless of entry layer —
a named deviation from mobility, which needed no such sharing.

AppState now exposes materialization_runtime (was materialization_service);
consumers (api/http/repo_roots.rs, tests) re-point to the runtime, which
delegates the purely-durable acquire_repo_root straight to the service.

Clears the domains/materialization/service.rs row from the DOMAIN_LIVE_VALVE
allowlist (11->10 files, 20->19 occurrences). service.rs shrank 625->419
lines (row deleted, under the 600 cap); app/mod.rs grew 693->694 (re-pinned).
