+++ scripts/max_lines_allowlist.txt
@@ -11,7 +11,6 @@ anyharness/crates/anyharness-contract/src/v1/events.rs 1092 existing AnyHarness 
 anyharness/crates/anyharness-contract/src/v1/sessions.rs 779 existing AnyHarness contract oversized file (+caller-selected idempotent create id)
 anyharness/crates/anyharness-lib/src/api/openapi.rs 692 existing AnyHarness oversized file (+hosting schema registration, +agent-auth state response, +gateway catalog endpoints) + enriched gateway-models schema fields (model-table §1) + workflow-runs + local materialization operations + workflow-run-workspaces paths/schemas + agent install progress schemas + worktree restore path/schemas + WorktreeBaseFetch + RuntimeDiskPressure schemas (central OpenAPI registry; per-route path/schema registration only)
 anyharness/crates/anyharness-lib/src/api/router_tests.rs 1898 existing AnyHarness oversized test file (+hosting PR-status route tests, +catalog version tests, +managed default-placement route test, +model-snapshot status route shape/auth tests, +agent-auth/install poke contract tests, +agent-auth/install poke-suppression proofs against the automatic_poke_engine seam (F-036 cycle-2 review item 1); the catalog-push handler tests are gone with the heartbeat transport)
-anyharness/crates/anyharness-lib/src/domains/materialization/service.rs 625 local materialization domain service (exact-ref workspace materialization + repo-root acquisition orchestration, in-process keyed operation-lock admission, idempotent replay/failure recording; blocking git mechanics split into acquire.rs) — split into service/ on next growth per guides/domains.md
 anyharness/crates/anyharness-lib/src/domains/workflows/service.rs 664 workflow durable rules (+run-control cancel-intent use case and helpers); split into service/ on next growth per guides/domains.md
 anyharness/crates/anyharness-lib/src/domains/workflows/service_tests.rs 1320 workflow-runs merge-gated domain suite (+2b partial-uniqueness contract on nonterminal session binding)
 anyharness/crates/anyharness-lib/src/api/workflow_runs_tests.rs 2790 workflow-runs HTTP + runtime/extension crossing suite (run-control PROOF-01 battery + 2b admission races: reservation/create/bind window, terminal release, PR1227-WORKSPACE-FENCE-01 purge/retire destruction-vs-late-bind proofs, and FENCE-02 bind->terminalize->destruction race proofs over the real executor)
@@ -74,7 +73,7 @@ apps/packages/product-client/src/domain/workspaces/cloud-work-inventory.test.ts 
 apps/packages/product-client/src/domain/sessions/activity.test.ts 660 relocated existing oversized domain test; no growth
 apps/packages/product-client/src/domain/sessions/intents/session-intent.test.ts 606 relocated existing oversized domain test; no growth
 anyharness/crates/anyharness-lib/src/api/router.rs 640 AnyHarness router (+goals/loops/activity/feed routes, +gateway catalog version route, +workflow-runs/cancel + local materialization operations + workflow-run-workspaces routes + worktree restore route)
-anyharness/crates/anyharness-lib/src/app/mod.rs 690 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +local materialization service wiring + worktree restore runtime wiring; -3: WorkspaceService construction/pass-through deleted, grid PR 8; net 0 in grid PR 9: WorkspaceRetireService joins AppState while the preflight-checker/purge/retire constructions move to the app/workspaces.rs wiring family)
+anyharness/crates/anyharness-lib/src/app/mod.rs 691 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +worktree restore runtime wiring; local materialization service wiring is now the materialization runtime valve, moved to app/materialization.rs per grid PR 7; -3: WorkspaceService construction/pass-through deleted, grid PR 8; net 0 in grid PR 9: WorkspaceRetireService joins AppState while the preflight-checker/purge/retire constructions move to the app/workspaces.rs wiring family)
 anyharness/crates/anyharness-lib/src/domains/agents/catalog/service_tests.rs 605 B-2 validation-cut rewrote the refusal assertions in place (single SESSION_MODEL_UNSUPPORTED); shrinks when the legacy gateway-models routes delete (C-track cutover)
 anyharness/crates/anyharness-lib/src/domains/loops/runtime.rs 867 loops runtime — native write path + emulated Codex scheduler (+2b derived-safe classification comment at the fire seam)
 anyharness/crates/anyharness-lib/src/live/sessions/actor/startup.rs 642 session startup (+goal/loop capability parsing, roster reconcile-on-attach)
